Update sniffer source for WoW-64. Added empty parser project.

+++ World of Warcraft/Source/Awps/Memory.cs
@@ -75,7 +75,7 @@ namespace Awps
 
                 Native.VirtualProtect(address, (uint)data.Length, 0x80, out oldProtect);
 
-                var realAddress = new IntPtr((uint)address);
+                var realAddress = new IntPtr((long)address);
 
                 Native.WriteProcessMemory(currentProcess.Handle, realAddress, data, data.Length);
